requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk
torch
transformers
fastapi
uvicorn
requests
sqlalchemy
psycopg[binary]
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-requests
bertopic
faiss-cpu
sentence-transformers
pytest
